requirements.txt
Flask==3.1.0
Pillow==11.3.0
moviepy==1.0.3
pydub==0.25.1
flask-cors==4.0.1
gunicorn==22.0.0
pillow-heif==0.17.0
